AuxiliaryTableau merged into Tableau.
Keeping two different types which supports almost the very same operations makes
the code too complicate. Now the auxiliary data is stored as a Maybe field in
Tableau record.

Now the function 'pivot' can operate on any tableau, ordinary one or auxiliary
one. In the latter case, it also updates the auxiliary data. The same for
functions 'chooseEnteringVariable' and 'chooseLeavingVariable'.