Added `makeTableau` and `makeTableau'` functions, to reduce coupling.
Till now the only way to make a `Tableau` was to build it directly with the
record syntax. This will be a problem if the structure of `Tableau` will change,
so the new functions should be used instead.

Some code still access the record directly, it will be refactored in the future.