Added new pivot operation that operates on the matrix form of tableau.
If the tableau is represented using a single matrix, the pivot operation is the
same as the Gaussian elimination's pivot, apart from the pivoting column. For
now tableaus are still represented in the old way, and the matrix form is
computed at each pivot, but this will change in future commits.